apps/backend/pyproject.toml
[project]
name = "bahu-erp-backend"
version = "0.1.0"
description = "Unified ERP backend"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "boto3>=1.35.0",
  "sqlalchemy>=2.0.32",
  "asyncpg>=0.29.0",
  "alembic>=1.13.2",
  "pydantic>=2.8.2",
  "pydantic-settings>=2.4.0",
  "python-jose[cryptography]>=3.3.0",
  "passlib[argon2]>=1.7.4",
  "python-multipart>=0.0.9",
  "celery>=5.4.0",
  "openpyxl>=3.1.5"
]

[tool.uv]
dev-dependencies = [
  "pytest>=8.3.2",
  "pytest-asyncio>=0.24.0",
  "httpx>=0.27.0",
  "ruff>=0.6.2",
  "mypy>=1.11.1"
]
